Add experimental support for multi systems launch and user configs

+++ Mod Files/ivistation/scripts/config/content_config.py
@@ -4,23 +4,45 @@ import xbmc
 import shutil
 import simplejson as json
 
-from abc import abstractmethod
 
 ROOT_DIR = xbmc.translatePath("Special://root/")
 EMULATORS_DIR = xbmc.translatePath("Special://root/ivistation/emulators/")
 CONFIGS_PATH = xbmc.translatePath("Special://root/ivistation/configs/")
 
 
-class BaseConfigManager(object):
-    def __init__(self):
-        pass
+class CoreConfigManager(object):
+    """
+    Loads a user declared configuration file, which affects all the roms for an emulator.
+    Note that this could be overriden by individual ROM configuration files that were declared by using
+    RomConfigManager, or that came with IVistation as ROM overrides.
 
-    @abstractmethod
-    def _get_config_file_path(self):
-        return ""
+    Although, overrides can be ignored.
 
-    def save_config(self, new_configuration, custom_path=None):
-        path = self._get_config_file_path() if custom_path is None else custom_path
+    In practice, this class tells IVistation what core to load for a system, and nothing more, as the
+    actual core configurations are handled by the core setting files.
+
+    Uses simple JSON structures.
+    """
+
+    def __init__(self, system):
+        super(CoreConfigManager, self).__init__()
+        self.system = system
+
+        self.path = os.path.join(
+            CONFIGS_PATH,
+            system,
+            "core.json"
+        )
+
+    def get_config(self):
+        if not os.path.isfile(self.path):
+            return None
+
+        with open(self.path, "r") as current_config:
+            return json.load(current_config)
+
+    def save_config(self, new_configuration):
+        path = self.path
 
         if not new_configuration:
             return
@@ -36,125 +58,94 @@ class BaseConfigManager(object):
             json.dump(new_configuration, current_config)
 
     def delete_config(self):
-        if os.path.isfile(self._get_config_file_path()):
-            os.remove(self._get_config_file_path())
+        if os.path.isfile(self.path):
+            os.remove(self.path)
 
 
-class RomConfigManager(BaseConfigManager):
+def get_core_path(emu_system, core):
     """
-    Loads a user declared configuration file, which affects only a given ROM from a given system.
+    Returns the full path to a core, for a given system.
     """
 
-    def __init__(self, rom_identifier, system):
-        super(RomConfigManager, self).__init__()
-        self.game_identifier = rom_identifier
-        self.system = system
-        self.override = False
+    possible_paths = (
+        os.path.join(EMULATORS_DIR, emu_system, core),
+        os.path.join(EMULATORS_DIR, "multi", core)
+    )
 
-        current_configs_path = os.path.join(CONFIGS_PATH, system)
+    for path in possible_paths:
+        if os.path.isdir(path):
+            return path
 
-        self.paths = [
-            os.path.join(
-                current_configs_path,
-                "{}.json".format(
-                    rom_identifier
-                )
-            ),
-            # Note how the override variant is later, is it's ignored if a user defined one exists
-            os.path.join(
-                current_configs_path,
-                "d {}.json".format(
-                    rom_identifier
-                )
+    raise KeyError("Core path not found, searched in: ", possible_paths, emu_system, core)
+
+
+def check_multicore_compatibility(emu_system, core_path):
+    """
+    Given the path to a core, (or the path to its executable),
+    checks if it's a multi-system core, and then checks if
+    the given system is supported.
+
+    If a regular core path is passed, and not a multi-system
+    core, False is returned.
+
+    If the info.json file is invalid, the whole core will
+    be deleted.
+
+    Args:
+        emu_system: The system to check compat for
+        core_path: The core path to check compat with
+    Returns:
+        boolean for validity
+    """
+
+    if core_path.endswith("default.xbe"):
+        core_path = os.path.dirname(core_path)
+
+    with open(os.path.join(core_path, "info.json"), "r") as ci:
+        core_info = json.load(ci)
+
+    # This is not a multi-system core, return false.
+    if "multi" not in core_info:
+        return False
+
+    if "systems" not in core_info["multi"]:
+        try:
+            shutil.rmtree(
+                os.path.dirname(core_path)
             )
-        ]
+        except Exception as e:
+            print("content_config.py: Couldn't delete ", core_path, emu_system, e)
+        return False
 
-    def _get_config_file_path(self):
-        return self.paths[0]
+    if emu_system not in core_info["multi"]["systems"]:
+        return False
 
-    def is_override(self):
-        return self.override
-
-    def get_config(self):
-        """
-        Returns the dictionary representing the configuration file, or None if it doesn't exist.
-        """
-
-        configuration = None
-
-        for config in self.paths:
-            if not os.path.isfile(config):
-                continue
-
-            with open(config, "r") as current_config:
-                configuration = json.load(current_config)
-
-            if os.path.basename(config).startswith("d "):
-                self.override = True
-
-        return configuration
-
-    def save_config(self, new_configuration, override=False):
-        super(RomConfigManager, self).save_config(
-            new_configuration,
-            custom_path=self.paths[1] if override else self.paths[0]
-        )
-
-
-class CoreConfigManager(BaseConfigManager):
-    """
-    Loads a user declared configuration file, which affects all the roms for an emulator.
-    Note that this could be overriden by individual ROM configuration files that were declared by using
-    RomConfigManager, or that come bundled with a certain emulator.
-
-    Although, emulator bundled configurations can be ignored.
-    """
-
-    def __init__(self, system):
-        super(CoreConfigManager, self).__init__()
-        self.system = system
-
-        current_configs_path = os.path.join(CONFIGS_PATH, system)
-
-        self.path = os.path.join(
-            current_configs_path,
-            "d  default.json"
-        )
-
-    def _get_config_file_path(self):
-        return self.path
-
-    def get_config(self):
-        if not os.path.isfile(self.path):
-            return None
-
-        with open(self.path, "r") as current_config:
-            return json.load(current_config)
+    return True
 
 
 def verify_emulator_existence(system, emulator):
     """
-    Verifies if an emulator really exists. True if true. Ba dum tss
+    Verifies if an emulator really exists. If the emulator isn't valid,
+    it will be deleted.
     """
 
-    if emulator is None or emulator == "" or emulator == " ":
+    if not isinstance(emulator, str) or emulator.strip() == "":
         return False
 
-    ver_emulator_dir = os.path.join(EMULATORS_DIR, system, emulator, "default.xbe")
+    try:
+        emu_exe = os.path.join(get_core_path(system, emulator), "default.xbe")
+    except KeyError:
+        return False
 
-    emu_exists = os.path.isfile(ver_emulator_dir)
+    emu_exists = os.path.isfile(emu_exe)
 
     if not emu_exists:
         try:
             shutil.rmtree(
-                os.path.join(
-                    EMULATORS_DIR,
-                    system,
-                    emulator
-                )
+                os.path.dirname(emu_exe)
             )
         except Exception as e:
-            print("content_config.py: Couldn't delete ", ver_emulator_dir, e)
+            print("content_config.py: Couldn't delete ", emu_exe, e)
 
     return emu_exists
 
@@ -170,45 +161,49 @@ def get_emulator_list_for_system(system):
 
     valid_systems = []
 
-    emulators_path = os.path.join(EMULATORS_DIR, system)
+    emulators_path = (
+        os.path.join(EMULATORS_DIR, system),
+        os.path.join(EMULATORS_DIR, "multi")
+    )
 
-    for emulator in os.listdir(emulators_path):
-        full_path = os.path.join(emulators_path, emulator)
+    for index in range(0, 2):
+        for emulator in os.listdir(emulators_path[index]):
+            full_path = os.path.join(emulators_path[index], emulator)
 
-        if os.path.isfile(os.path.join(full_path, "default.xbe")):
-            valid_systems.append(emulator)
-        else:
-            try:
-                shutil.rmtree(
-                    full_path
-                )
-            except Exception as e:
-                print("content_config.py: Couldn't delete ", full_path, e)
+            # Multi-systems core
+            if index == 1 and not check_multicore_compatibility(system, full_path):
+                continue
+
+            # Standalone systems core
+            if os.path.isfile(os.path.join(full_path, "default.xbe")):
+                valid_systems.append(emulator)
+            else:
+                try:
+                    shutil.rmtree(
+                        full_path
+                    )
+                except Exception as e:
+                    print("content_config.py: Couldn't delete ", full_path, e)
 
     return valid_systems
 
 
 def get_system_defaults():
     """
-    Return the system bundled emulators, which cannot be removed by the user, by default.
+    Return the system bundled emulators, which cannot be removed by the user.
     Read only.
     """
+
     with open(os.path.join(ROOT_DIR, "ivistation\\data\\default_emus.json"), "r") as defaults:
         return json.load(defaults)
 
 
-def get_core_path(emu_system, core):
-    return os.path.join(EMULATORS_DIR, emu_system, core)
-
-
 def get_core_info(emu_system, core_id):
-    core_path = get_core_path(emu_system, core_id)
-
-    with open(os.path.join(core_path, "info.json"), "r") as ci:
+    with open(os.path.join(get_core_path(emu_system, core_id), "info.json"), "r") as ci:
         return json.load(ci)
 
 
-def is_valid_ini_line(line):
+def _is_valid_ini_line(line):
     # Regular expression pattern for matching section (header)
     section_pattern = re.compile(r'^\s*\[([^]]+)]\s*$')
 
@@ -235,7 +230,7 @@ def prepare_config_file(config_path):
         config_file.seek(0)
 
         for line in lines:
-            if is_valid_ini_line(line):
+            if _is_valid_ini_line(line):
                 config_file.write(line)
 
         config_file.truncate()
